gemini: rootdir: Tweak VM on boot
Remove CAF tweaks to nonexistent modules on our kernel.

Don't modify swappiness so it will be left at 60 which is kernel default.
It prioritizes file pages to be swapped first vs anonymous pages, this shoud boost apps performance.

Lower vfs_cache_pressure to 80 so more file pages will be kept in RAM which will finely scale with our
new swappiness level which should result in overall better caching without costly calls to UFS.

Remove allocstall_threshold disablement because it seems nice to me and hopefully should result in
better LMKD choices. Kernel sets it to 70 as default.

Remove readahead tuning as it is forced to 0 anyways.

Signed-off-by: xNombre <kartapolska@gmail.com>
Change-Id: Icf7486cad6bdd254180b3b77fc5d17f034ec4dad
Signed-off-by: iKeramat <ikeramat80@gmail.com>